Rename the variables with more precise names and extract common types
