+++ packages/graph/src/index.ts
@@ -1,4 +1,5 @@
 export * from './components/Foo';
+export * from './common/types';
 export * from './layers/ResponsiveLayer';
 export * from './layers/DataLayer';
 export * from './layers/AxisLayer';
